detect/file-data: move tests into tests/
